Remove payments tab
